Feat:架构优化,从主文件中剥离PyProject和SuffixProject,易于维护
