+++ src/autocoder_nano/project/__init__.py
@@ -1,0 +1,4 @@
+from autocoder_nano.project.pyproject import PyProject
+from autocoder_nano.project.suffixproject import SuffixProject
+
+__all__ = ["PyProject", "SuffixProject"]
