+++ src/autocoder_nano/version.py
@@ -1,3 +1,3 @@
-__version__ = "0.1.13"
+__version__ = "0.1.14"
 __author__ = "moofs"
 __license__ = "Apache License 2.0"
